Update tab 2,3

+++ eoqLab/Models/StockParams.cs
@@ -13,7 +13,7 @@
         public string UnitName { get; set; }
         public int BrandID { get; set; }
         public string BrandName { get; set; }
-        public float Amount { get; set; }
-        public float Price { get; set; }
+        public decimal Amount { get; set; }
+        public decimal Price { get; set; }
     }
 }
